requirements.txt
python-telegram-bot==20.7
openpyxl==3.1.2
python-dotenv==1.0.0
matplotlib==3.8.4
reportlab==4.1.0

